+++ scripts/build_sc_machine.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -eo pipefail
-
-if [[ -z ${SC_MACHINE_PATH+1} ]];
-then
-  source "$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)"/set_vars.sh
-fi
-
-"${SC_MACHINE_PATH}/scripts/build_sc_machine.sh"