+++ scripts/run_sc_web.sh
@@ -1,7 +1,7 @@
-#!/bin/bash
-set -e
+#!/usr/bin/env bash
+set -eo pipefail
 
-if [[ -z ${SC_WEB_PATH+1} ]];
+if [ -z "${SC_WEB_PATH}" ];
 then
   source "$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)"/set_vars.sh
 fi