+++ scripts/build_kb.sh
@@ -1,2 +1,9 @@
-#!/bin/bash
-../sc-machine/bin/sc-builder -f --clear -c ../ostis-web-platform.ini -i ../repo.path -o ../kb.bin
+#!/usr/bin/env bash
+set -eo pipefail
+
+if [ -z "${SC_MACHINE_PATH}" ];
+then
+  source "$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)"/set_vars.sh
+fi
+
+"${SC_MACHINE_PATH}/scripts/build_kb.sh" "$@"