+++ scripts/build_sc_web.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
-if [ -z "${SC_MACHINE_PATH}" ];
+if [ -z "${SC_WEB_PATH}" ];
 then
   source "$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)"/set_vars.sh
 fi
 
-"${SC_MACHINE_PATH}/scripts/run_sc_server.sh" "$@"
+"${SC_WEB_PATH}/scripts/build_sc_web.sh" "$@"