+++ scripts/clean.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-rm -rf ../sc-machine
-rm -rf ../sc-web
-rm -rf ../kb.bin
-rm -rf ../geometry.drawings