+++ scripts/formats.sh
@@ -1,0 +1,4 @@
+stage()
+{
+  tput setaf 4 && tput bold && echo "$1" && tput sgr0
+}