+++ docs/changelog.md
@@ -13,14 +13,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use py-sc-client and ts-sc-client for sc-server instead of sctp-clients for sctp-server
 
 ### Added
+- Generalize all scripts
 - Configure events and agents threads
 - Add latex documentation from OSTIS Standard
 - Add CI
 
 ### Changed
+- Link scripts with sc-machine and sc-web scripts
+- Deprecate `run_scweb.sh`, add `run_sc_web.sh` instead
+- Deprecate and divide `prepare.sh` into `install_submodules.sh`, `install_dependencies.sh`, `build_sc_machine.sh` and `build_sc_web.sh`
 - Unify config file for sc-server and sc-builder
 
 ### Removed
+- Remove windows scripts
+- Remove drawings and scp scripts
+- Remove dump scripts
 - Remove sctp-server scripts
 
 ## [0.6.1] - 28.04.2022